requirements.txt
websockets
asyncio
psycopg2-binary
